adds validation

+++ spring-mvc/src/main/java/ch/fhnw/webfr/sivakumm/zirper/web/ZirpController.java
@@ -3,11 +3,14 @@ package ch.fhnw.webfr.sivakumm.zirper.web;
 import java.util.Date;
 import java.util.Optional;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -43,7 +46,10 @@ public class ZirpController {
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public String createZirp(Zirp zirp) {
+    public String createZirp(Model model, @Valid Zirp zirp, BindingResult result) {
+        if (result.hasErrors()) {
+            return "redirect:/zirps";
+        }
         zirp.setDate(new Date());
         zirpRepository.save(zirp);
         return "redirect:/zirps";
@@ -51,12 +57,26 @@ public class ZirpController {
 
     @RequestMapping(value = "/{id}", params = "update", method = RequestMethod.GET)
     public String getUpdatePage(@PathVariable String id, Model model) {
-        model.addAttribute("zirp", zirpRepository.findById(id).get());
+        Optional<Zirp> zirp = zirpRepository.findById(id);
+        if (zirp.isEmpty()) {
+            return "404";
+        }
+
+        model.addAttribute("zirp", zirp.get());
         return "zirp/edit";
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-    public String updateZirp(@PathVariable String id, Zirp zirp) {
+    public String updateZirp(@PathVariable String id, @Valid Zirp zirp, BindingResult result) {
+        if (result.hasErrors()) {
+            return "zirp/edit";
+        }
+
+        Optional<Zirp> original = zirpRepository.findById(id);
+        if (original.isEmpty()) {
+            return "404";
+        }
+
         zirp.setId(id);
         zirpRepository.save(zirp);
         return "redirect:/zirps/" + id;
@@ -64,6 +84,10 @@ public class ZirpController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public String deleteZirp(@PathVariable String id) {
+        if (zirpRepository.findById(id).isEmpty()) {
+            return "404";
+        }
+
         zirpRepository.deleteById(id);
         return "redirect:/zirps";
     }
